Move list_index into separate file
