+++ libvast/CMakeLists.txt
@@ -121,6 +121,7 @@ set(libvast_sources
     src/ids.cpp
     src/index/address_index.cpp
     src/index/enumeration_index.cpp
+    src/index/list_index.cpp
     src/index/port_index.cpp
     src/index/string_index.cpp
     src/index/subnet_index.cpp
@@ -345,6 +346,7 @@ set(tests
     test/index/address_index.cpp
     test/index/arithmetic_index.cpp
     test/index/enumeration_index.cpp
+    test/index/list_index.cpp
     test/index/port_index.cpp
     test/index/string_index.cpp
     test/index/subnet_index.cpp
