+++ libvast/src/index/address_index.cpp
@@ -13,6 +13,7 @@
 
 #include "vast/index/address_index.hpp"
 
+#include "vast/detail/overload.hpp"
 #include "vast/index/container_lookup.hpp"
 #include "vast/type.hpp"
