+++ libvast/src/index/string_index.cpp
@@ -15,6 +15,7 @@
 
 #include "vast/bitmap_algorithms.hpp"
 #include "vast/defaults.hpp"
+#include "vast/detail/overload.hpp"
 #include "vast/index/container_lookup.hpp"
 #include "vast/type.hpp"
