+++ libvast/src/index/subnet_index.cpp
@@ -13,6 +13,7 @@
 
 #include "vast/index/subnet_index.hpp"
 
+#include "vast/detail/overload.hpp"
 #include "vast/index/container_lookup.hpp"
 #include "vast/type.hpp"
