+++ libvast/src/system/indexer.cpp
@@ -30,6 +30,7 @@
 #include "vast/system/report.hpp"
 #include "vast/table_slice_column.hpp"
 #include "vast/value_index.hpp"
+#include "vast/value_index_factory.hpp"
 #include "vast/view.hpp"
 
 #include <caf/attach_stream_sink.hpp>
