+++ libvast/src/value_index.cpp
@@ -13,15 +13,12 @@
 
 #include "vast/value_index.hpp"
 
-#include "vast/base.hpp"
-#include "vast/defaults.hpp"
-#include "vast/index/container_lookup.hpp"
+#include "vast/value_index_factory.hpp"
 
-#include <cmath>
+#include <caf/deserializer.hpp>
+#include <caf/serializer.hpp>
 
 namespace vast {
-
-// -- value_index --------------------------------------------------------------
 
 value_index::value_index(vast::type t, caf::settings opts)
   : type_{std::move(t)}, opts_{std::move(opts)} {
@@ -146,82 +143,6 @@ caf::error inspect(caf::deserializer& source, value_index_ptr& x) {
   if (x == nullptr)
     return make_error(ec::unspecified, "failed to construct value index");
   return x->deserialize(source);
-}
-
-// -- list_index -----------------------------------------------------------
-
-list_index::list_index(vast::type t, caf::settings opts)
-  : value_index{std::move(t), std::move(opts)} {
-  max_size_ = caf::get_or(options(), "max-size",
-                          defaults::index::max_container_elements);
-  auto f = detail::overload{
-    [](const auto&) { return vast::type{}; },
-    [](const list_type& x) { return x.value_type; },
-  };
-  value_type_ = caf::visit(f, value_index::type());
-  VAST_ASSERT(!caf::holds_alternative<none_type>(value_type_));
-  size_t components = std::log10(max_size_);
-  if (max_size_ % 10 != 0)
-    ++components;
-  size_ = size_bitmap_index{base::uniform(10, components)};
-}
-
-caf::error list_index::serialize(caf::serializer& sink) const {
-  return caf::error::eval(
-    [&] { return value_index::serialize(sink); },
-    [&] { return sink(elements_, size_, max_size_, value_type_); });
-}
-
-caf::error list_index::deserialize(caf::deserializer& source) {
-  return caf::error::eval(
-    [&] { return value_index::deserialize(source); },
-    [&] { return source(elements_, size_, max_size_, value_type_); });
-}
-
-bool list_index::append_impl(data_view x, id pos) {
-  auto f = [&](const auto& v) {
-    using view_type = std::decay_t<decltype(v)>;
-    if constexpr (std::is_same_v<view_type, view<list>>) {
-      auto seq_size = std::min(v->size(), max_size_);
-      if (seq_size > elements_.size()) {
-        auto old = elements_.size();
-        elements_.resize(seq_size);
-        for (auto i = old; i < elements_.size(); ++i) {
-          elements_[i] = factory<value_index>::make(value_type_, options());
-          VAST_ASSERT(elements_[i]);
-        }
-      }
-      auto x = v->begin();
-      for (auto i = 0u; i < seq_size; ++i)
-        elements_[i]->append(*x++, pos);
-      size_.skip(pos - size_.size());
-      size_.append(seq_size);
-      return true;
-    }
-    return false;
-  };
-  return caf::visit(f, x);
-}
-
-caf::expected<ids>
-list_index::lookup_impl(relational_operator op, data_view x) const {
-  if (!(op == ni || op == not_ni))
-    return make_error(ec::unsupported_operator, op);
-  if (elements_.empty())
-    return ids{};
-  auto result = elements_[0]->lookup(equal, x);
-  if (!result)
-    return result;
-  for (auto i = 1u; i < elements_.size(); ++i) {
-    auto mbm = elements_[i]->lookup(equal, x);
-    if (mbm)
-      *result |= *mbm;
-    else
-      return mbm;
-  }
-  if (op == not_ni)
-    result->flip();
-  return result;
 }
 
 } // namespace vast
