+++ libvast/test/index/string_index.cpp
@@ -20,9 +20,11 @@
 
 #include "vast/concept/printable/to_string.hpp"
 #include "vast/concept/printable/vast/bitmap.hpp"
+#include "vast/detail/overload.hpp"
 #include "vast/load.hpp"
 #include "vast/save.hpp"
 #include "vast/table_slice.hpp"
+#include "vast/value_index_factory.hpp"
 
 #include <caf/test/dsl.hpp>
