+++ libvast/vast/hash_index.hpp
@@ -13,10 +13,12 @@
 
 #pragma once
 
+#include "vast/byte.hpp"
 #include "vast/concept/hashable/uhash.hpp"
 #include "vast/concept/hashable/xxhash.hpp"
 #include "vast/data.hpp"
 #include "vast/detail/assert.hpp"
+#include "vast/detail/overload.hpp"
 #include "vast/detail/stable_map.hpp"
 #include "vast/detail/type_traits.hpp"
 #include "vast/value_index.hpp"
